fix: fix styles and logic on Game screen
